Allow for format to be embedded in -f parameter or CPMTOOLSFMT.
In addition to accepting a format name via -f or CPMTOOLSFMT,
this patch allows for the actual format text to be used, avoiding
the need to edit diskdefs for every format. This is particularly
handy when disk formats, such as SASI drives, may be customized
(example, Magnolia Microsystems SASI drive partitioning format).
A user may then create a wrapper for cpmtools that reads format
information off the SASI image, creates a diskdef for it, and passes
that on the commandline or environment to the cpmtools commands.

Pull-request: #3 #1
Signed-off-by: Stephan Linz <linz@li-pro.net>