Fix problem where Kaypro disk reserved space was clobbered.
This issue affects any format where the vendors DPB.ALV0/1 allocates
more blocks than what is implied by DPB.DRM (maxdir). Kaypro did this
in order to extend the boot track to handle the larger system size.

This fix is to add a 'dirblks' parameter that essentially handles
the extra blocks reserved by DPB.ALV0/1. In the case of the Kaypro,
the maxdir parameter of 64, with a 2K block size, resulted in one block
reserved for the directory. When files were added to the image, cpmtools
would allocate the next block after the directory and thus wipe out the
boot code stored there.

This new parameter is optional, and the value will be computed based
on maxdir if it is left out.

The Kaypro diskdefs now reserve this extra block so that the boot code
is protected.